見た目の修正とユーザー管理

+++ app/controllers/calendars_controller.rb
@@ -1,5 +1,7 @@
 class CalendarsController < ApplicationController
   before_action :authenticate_user!
+  before_action :set_event, only: [:edit, :update] 
+
   def index
     @events = Event.all 
   end
@@ -18,15 +20,14 @@ class CalendarsController < ApplicationController
   end
 
   def edit
-    @event = Event.find(params[:id])
     unless @event.user_id == current_user.id
-      redirect_to action: :index
+      redirect_to root_path
     end
   end
 
   def update
-    @event = Event.find(params[:id])
-    if @event.update(event_parameter)
+    @event.update(event_parameter)
+    if @event.save
       redirect_to action: :index
     else
       render :edit
@@ -35,8 +36,12 @@ class CalendarsController < ApplicationController
 
   def destroy
     event = Event.find(params[:id])
-    event.destroy
-    redirect_to action: :index
+    if current_user.id == event.user_id
+      event.destroy
+      redirect_to action: :index
+    else
+      redirect_to root_path
+    end
   end
 
   private
@@ -44,4 +49,9 @@ class CalendarsController < ApplicationController
   def event_parameter
     params.require(:event).permit(:title, :content, :start_time).merge(user_id: current_user.id)
   end
+
+  def set_event
+    @event = Event.find(params[:id])
+  end
+
 end
